#511 add data monitoring menu in lnb
