+++ discovery-frontend/src/app/monitoring/audit/job-log/component/job-detail/job-detail.component.ts
@@ -24,6 +24,7 @@ import { isUndefined } from 'util';
 import { LogEditorComponent } from '../../../component/log-editor/log-editor.component';
 import { MomentDatePipe } from '../../../../../common/pipe/moment.date.pipe';
 import { ActivatedRoute } from '@angular/router';
+import { CommonUtil } from '../../../../../common/util/common.util';
 
 
 @Component({
@@ -103,6 +104,7 @@ export class JobDetailComponent extends AbstractPopupComponent {
   /*-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=
   | Public Method
   |-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=*/
+  public convertMilliseconds:Function = CommonUtil.convertMilliseconds;
 
   /**
    * 상세 팝업 종료
@@ -124,25 +126,6 @@ export class JobDetailComponent extends AbstractPopupComponent {
    */
   public isPlanNull(): boolean {
     return isUndefined(this.audit.plan) || this.audit.plan === null;
-  }
-
-  /**
-   * ms를 min과 sec로 변환
-   * @param {number} ms
-   * @returns {string}
-   */
-  public convertMilliseconds(ms: number): string {
-    if (ms === undefined) {
-      return 0 + ' sec';
-    }
-
-    if (ms < 1000) {
-      return ms + 'ms'
-    }
-
-    const min = Math.floor((ms / 1000 / 60) << 0);
-    const sec = Math.floor((ms / 1000) % 60);
-    return min !== 0 ? (min + ' min ' + sec + ' sec') : (sec + ' sec');
   }
 
   /*-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=
