+++ discovery-frontend/src/app/monitoring/audit/job-log/job-log.component.html
@@ -27,9 +27,6 @@
       <li class="ddp-selected">
         <a [routerLink]="['/management/monitoring/audit']">{{'msg.comm.menu.manage.monitoring.audit' | translate}}</a>
       </li>
-      <li>
-        <a [routerLink]="['/management/monitoring/lineage']">{{'msg.comm.menu.manage.monitoring.data-lineage' | translate}}</a>
-      </li>
     </ul>
     <!-- //tab -->
   </div>
