+++ discovery-frontend/src/app/monitoring/audit/log-statistics/log-statistics.component.html
@@ -28,9 +28,6 @@
       <li>
         <a [routerLink]="['/management/monitoring/audit']">{{'msg.comm.menu.manage.monitoring.audit' | translate}}</a>
       </li>
-      <li>
-        <a [routerLink]="['/management/monitoring/lineage']">{{'msg.comm.menu.manage.monitoring.data-lineage' | translate}}</a>
-      </li>
     </ul>
     <!-- //tab -->
